[TASK] Prevent Caching-Disabling if nonces are set
In our PoC we want cache the page as usual, and replace the nonce (which is invalid after the very first use) with a current one directly from the webserver using SSI with StaticFileCache. 

To achieve this, entries in die INTincScript-Array must be prevented, as these flags the page as non-cachable at all.